Install Symfony 3.3
Our mission is to use Symfony 3.4 as this will be the next LTS version of the software. This version will be released in November. For now we use the current stable release. Which is version 3.3.

To get Symfony 3.3.9 an update of twig/twig was also required.

In addition the monolog-bundle was also upgraded.

Also fixed the BC breaking config change that was introduced in version 3.3. It requires the move of the trusted proxies config entry to live in the front controller. More on this in this blogpost: https://symfony.com/blog/fixing-the-trusted-proxies-configuration-for-symfony-3-3

As the trusted proxies have not yet been installed, this means no changes to the app.php and app_dev.php need to be made.