Install Doctrine Migrations Bundle 1.2
Changes to the database entities will be tracked using Doctrine
Migrations. These migration provide a means to update but also rollback
database schema changes.

This installation not only installs but also configures the existing
Symfony installation with the Doctrine Migrations Bundle. Also adding
the location where the migrations will be stored (/app/DoctrineMigrations)

A first entry in the readme was added, instructing the developer to also
install the doctrine migrations during installation of the application.