Install Symfony Assetic Bundle 2.8.2
Assetic will be used to manage assets in the application. This is a
minimal installation, and might need some tweaks. At the moment of
writing this, we do not yet know all the details about the design
decisions.

Not only is the bundle installed, it is also configured into the
application.