Add Contact and Supplier entities
Contacts are the users of the system that log in for a given supplier.
The data of these contacts are tracked for Surfnet to be able to contact
these users.

This data might later be retrieved directly from the Teams api. In that
case all team members for the Suppliers team will be listed eliminating
the need to keep track of these in the application.

Basic information about these entities have been added to
the ubiquitous_language.md file.