Improve styling of the CreateSupplier form
The form styling is taken from Teams NG. The styling does not correspond
for the full 100% but this is a nice starting point.

Further modifications to the Symfony form templates are required to tap
into all the form capabilities available to us.

In addition the overview endpoint was removed from the Supplier
controller. This overview will never be created, instead the form
redirects to the / route of the application.