Change flash message behaviour
The flash message provided by the Teams NG styling is not meant to be
used for showing form validation errors. Changes to the flash message
styling was reverted.

The additional form validation messages (originating from the Handler)
are now displayed as being a message. This class is not yet decorated
in a stylesheet.