Change the datatype of Services id field
The need to generate id's before creating a Service is introduced by
the use of the CQRS priciple. As a command can never return data (for
example the Id of the service that was just created). Using Uuids as
primary keys allows us to create primary keys in a safe manner. Before
the record is created.

This change includes the schema change (migration) and the change in
the entity model.

Running the migration is a simple as running:

<pre>
bin/console doctrine:migrations:migrate
</pre>