Implement the CreateServiceCommand
To get the CreateServiceCommand to work quite some work had to be done.

1. A Handler was created to handle the command. This handler requires
the ServiceRepository to be able to save the command.

2. The ServiceController is responsible for creating the
CreateServiceCommand. To do so, additional dependencies are required.

First of all the SamlServiceService is used to generate a new id for the
Service entity that is about to be created. This entity id is later used
to load the form in edit mode.

Secondly the AdminSwitcherService is used to load the Supplier from the
currently active session.

Finally the TicketService is used to load a ticket number for the
service. For now this is a random Uuid.

3. When the Handler finishes handling it's command, the Controller
redirects to the editService action on the same Controller.

An unit and web test where provided to validate the correct
addition of the Service.

Services are registered for the production and test environments.