Add the Service form
1. The Service is loaded from the SamlServiceService
This required changes to the ServiceRepository. These changes have been
implemented in the Infrastructure repository and in the in memory test
repository.

2. Create the EditServiceCommand
The command is created by the ServiceCommandFactory, based on the Entity
it is fed with.

3. Create the EditServiceType (form)
The form is created using the EditServiceCommand. The form is made up
of two other FormTypes (Attribute, Contact).