Add and configure tinymce for the translation UI
Tinymce integration is identical to the old registration form, with
bonus: style dropdown in the toolbar.

The old registration form uses a custom Symfony form type to enable
tinymce on the textareas. The new version of the lexik bundle only
allows 'text' and 'textarea' form types so we need a different
approach. This commit overrides the layout template to initialize
tinymce and adds click events to the edit and save buttons to
initialize and remove tinymce on the textarea.