Create and test the Fetcher
The Fetcher is responsible for loading the Metadata Xml from an Url.
This url is set on the client. The client then uses Guzzle to load the
xml from given page. If the response is invalid, an exception is thrown.

The Fetcher is registered in the services.yml file as the configuration
of the Fetcher requires some additional configuration that is not
autowireable.

The parameters.yml.dist file was updated with a sensible default value
for request timeouts. This value was taken from the SP registration form
project.