Integrate the LoadMetadataCommand in Controller
The CommandHandler for the LoadMetadataCommand will utilize the Fetcher
to read the Xml from the ImportUrl.

An exception is thrown if neither ImportUrl or Xml input is provided in
the form but the import button was clicked. This should later be handled
in a more elegant way, for example by showing a validation message.