Create and test the Parser
The Parser is responsible for parsing the Metadata Xml. This XML was
previously loaded by the Fetcher. Or directly injected from the web
form.

In this commit all dependencies of the Parser are added. This is a copy
paste from the previous SP Form. Some dependencies have been contracted
in interfaces.

Some constants that where loaded from the SimpleSamlPHP library have
been imported to the Parser itself. This to prevent adding not needed
third party libraries.