Fix the issues with the SupplierController
The supplier controller not yet used the Supplier service to load the
currently active supplier. It was using the AdminSwitcherService for
this task. It was relieved of that specific task in a previous refactor.

To fix the problem, the SupplierService was injected on the controller.
And the supplier is now fetched from said service, fixing the problem.