Add the pastedMetadata field to Service
This field will contain the metadata that is pasted into the form.
Previously the metadataXml field was used for this purpose. This is in
conflict with the inner working of the Metadata generator that is to be
created in the next feature.

This refactor also introduced a schema change. This change is captured
in the migration.