Add ManageClient for reading and writing to Manage
The Client uses an HttpClient for the actual communication with the
API. A Service entity is passed to the client (an implementation of the
PublishServiceRepository). The Service entity is transformed to a
PublishRequest which is used to transform the XML of the entity to a
Json encoded string. This Json is then posted to the Manage API.