Integrate the PublishServiceClient
The Client is integrated utilising the PublishServiceCommand and its
handler.

The command is triggered when the user presses the publish button.

The command for now is only covered by a web test. An integration test
could/should be added.

The client is configured in the services.yml and additional manage
client api parameters have been added to the parameters.yml.dist file.