Introduce SP Registration validation to dashboard
The custom validation has been reintroduced and is perfomred on publish.
Validation is not triggered on save.

The Validators have not been fully integrated into the application as
some validators require janus interaction. This should be replaced with
manage interaction.

Introduce SP Registration validation to dashboard

The custom validation has been reintroduced and is perfomred on publish.
Validation is not triggered on save.

The Validators have not been fully integrated into the application as
some validators require janus interaction. This should be replaced with
manage interaction.