Publish and push metadata to manage (part II)
To get data into manage, metadataXml is rendered just in time. And is
then published to manage. This entailed adding the Generator as a
dependency on the PublishEntityCommandHandler. This also required some
refactoring of it's tests.

After publishing the push to EB is still posted to the endpoint.

A web test makes sure the publishing of the entity works correctly
from an end-to-end perspective. The web test tests the happy flow
and the flow when validation fails.