Removed the ValidMetadata validator
The valid metadata validation is no longer required as we want to allow
reading of pasted metadata (based of the metadata of an unreadable
metadata endpoint).

This ended in removing the metadata tests and fixture data. During
cleaning these files up, I also boyscouted the ValidLogoValidatorTest,
removing an unnecessary whitespace.