Add the PrivacyQuestion form
The form can be used in two modes, being Create and edit mode. Both use
the same form type and command. This command is to be fed to a command
handler who can distinct between the two modes.

The webtests have been updated to test if the rendered action contains
form fields.