Add the Privacy question toggle to the form
In doing so, the toggle option is added to the Service entity. This
field is called: privacyQuestionsEnabled and is set to enabled by
default. The schema change can be applied using the Doctrine migration
that was also added.

The option is added as a checkbox to the create and edit forms for the
Service type.

When loading the Service edit form, the current state of the checkbox
is reflected on the form.

The Command handlers for the Service forms handle the toggle. The
correct inner working is validated in the integration tests.