Send the PrivacyQuestion Metadata to manage
The PublishEntityCommandHandler received the AttributesMetadata-
Repository. This repository is used to build the Privacy question (PQ)
metadata (and the rest of the augmented metadata later on).

A factory was added, and tasked with converting the PQ data from the
attribute repository into data suitable for putting into Manage.

Some boyscouting (typehint fixes) was done on the Service,
PQ and PrivacyQuestionsCommand.