Implement editing entities in test
It is now possible to edit entities which are published to the test
(connect) environment in manage. When a users clicks 'edit' on a
published entity in the service list, the user is redirected to an
existing draft if one exists. Otherwise:

 - the entity is retrieved from manage
 - the metadata is retrieved from manage
 - the dashboard checks if the entity belongs to the currently active service
 - a new draft entity is created
 - the metadata is imported
 - some extra fields are set on the new entity (like attribute
   motivation)
 - the user is redirected to the edit form of the new entity

This commit removes the 'state' on the 'Entity' ORM entity. Entities
stored in the dashboard are always 'draft' entities. Entities retrieved
from manage are always 'published' entities. We might have to revise
that when we implement 'publish to production'.

The janus ID field on an entity is renamed to manage ID. We store the
internal manage ID there so we can determine to what manage entity a
draft belongs (if any), and if we should POST or PUT the new metadata.