Create the MailMessageFactory
The MailMessageFactory builds Messages that can be set on a MailCommand.
The factory is injected with the translator (to translate the subjects
of messages) and the template engine. In addition the mail addresses
are also injected (to, from, noreply)

The factory is configured in the service container and the parameters
required to create it are added to the parameters template files.

A rather primitive test was added to test the factory.