Rename PublishEntity to reflect test intention
For now the Publishing of entities where always aimed at the test env.
In the upcomming changes, the entities can also be published to
production. To facilitate this change, the current name of the publish
command/handler and tests where misleading. And have been change to
more clearly indicate the test environment is being targeted.