Remember logo width and height and send it to manage
Instead of calculating the logo dimensions every time we load
metadata, only do it when the dimensions are unknown. This way the
original width and height won't get lost. We also throw an error if
the dimensions could not be determined: the width and height are not
optional.

Manage only accepts the logo if the dimensions are set.

Style change included: use 'mdui' instead of 'ui' as XML namespace.