Allow privileged users to create draft entities for production
SURF-administrators can enable service administrators to create
entities on the production environment without going to test
first.

When the checkbox 'allow production entities' is enabled, users are
allowed to create entities directly for the production environment. To
distinguish creating entities for test and production, the navigation
link 'Add entity' is removed and the 'Add' actions are shown above the
entity list.