Improve logo validation
Logo validation was previously only performed using the 'getimagesize'
function. This function internally resolved the logo and returned either
boolean false (not an image) or an array with the image details.

This was not sufficient as some logos seem to be valid PNG/GIF but are
not available for other reasons. By adding an extra curl based download,
we gain more insights in the how and why an image might be unavailable.

Extra logging was also added to enable us to debug logo failures more
easily in the future.