Support EntitiesDescriptor during import
While importing metadata we did not yet support the occurence of
multiple EntitiesDescriptor. This concept allows an SP/IdP to describe
multiple entities in a single metadata file.

For now we only allow the use of EntitiesDescriptor when only a single
entity is described in the document. When more than one documents exist,
an error message is shown, informing the user we do not support this
format.

Two web tests where added to cover these two new situations.