Setup of Parsley in project
A jquery resolution configuration was required in the package.json to
prevent installation of multiple jquery version in the project. See the
package.json for more details.

In addition, the require statements found in the application.js have
been replaced with import statements. And validation.js was added to the
list of imports.