Set translatable attr motivation warning message
The message can be translated in the messages translation domain as the
translator used in twig templates are in that specific domain.

The message is set on a data attribute of the motivation fields. And
could potentially contain per attribute specific translations. For now
a generic text is set.

The attributes Twig block (from the Symfony div form layout) was
overloaded to instruct this block to translate said data attribute.