Re-enable running encore in production mode
Thanks to an explanation found on the webpack encore issues page:

> Lyrkan: on Aug 15, 2017
>
> So, looking back at it a bit more, the version of Webpack shipped with
> Encore currently installs uglifyjs-webpack-plugin v0.4.6 which uses
> uglify-js v2.8.29. This version of uglify-js doesn't support ES6 and,
> since vendors don't go through Babel it fails if you require something
> from them written using ES6. It works using the files in dist since
> they have already been processed by Babel using the ES2015 preset.

SO the solution to our problem was not to include Parsley from the 'src'
folder, but to read it from the ES2015 compliant 'dist' folder.

The makeRelease.sh was also updated to run encore in production mode,
reverting the previously made change.