Don't overwrite fields when updating existing entities
The problem was two-fold:

1. When publising an entity already existing in manage, entity fields
such as 'active' and 'allowall' were always reset to their original
value. This commit changes the behaviour to only send the values on
the initial push to manage, and not on subsequent pushes.

2. The merge-write implementation was sending a list of metadata
fields, causing all metadata fields not in that list to be
deleted. This has been solved by using the flat format so all fields
not known to the dashboard are untouched.