Introduce a word-threshold configuration option
TextTypes with the 'keep talking' feature can now also be set with an
optional word-threshold attribute. Should be set with an integer value.
When the attribute is omitted the default (set to 10) word threshold
will be used.

The attribute motivation TextTypes are now configured to set a three
word input warning threshold.