Update commands and handlers
The create and edit service commands and handlers have been updated
to reflect the new way of working with the privacy questions answered
status.

The status is still tracked on the edit command for display purpose on
the symfony form. But the state is no longer persisted in the entity.