Specify manifest.json
Encore in production mode now creates versioned js/css files to prevent
unwarranted caching at the client side. This however requires symfony
asset manager to be aware of which version to use. Conveniently encore
now dumps a manifest.json, which the asset manager can use to determine
which version of the assets should be used.

For more information read:
https://symfony.com/doc/3.4/frontend/encore/versioning.html#loading-assets-from-entrypoints-json-manifest-json