Integrate the jquery.modal
The choose entity type form is now displayed in a modal. Some styling
was added to display the modal in a more material design'ish' style.
This style is based on the other SURFnet applications.

The change to list.html.twig includes the addition of two divs,
containing the modal contents. Twig is used to render the contents
of the entity_type action. The triggering of the modal is done by
updating the existing 'create entity' links. The anchors have been
changed to target the modal divs, and a rel attribute indicates the
role of the link (it should open the targeted modal).

The change to EntityCreateController was required to update the form
actions of the choose entity type forms. These need to be set explicitly
in order to handle the forms in the correct controller action.