Merge EntityDto change with Contact requirement
The RequestRemovalCommandHandler requires the contact to be present
in order to be able to report who requested the delete request.

The solution chosen was to add the contact on the DTO an piggy back it
into the command handler that way. Not the most elegant of solutions.
I've thought about DI'ing the contact on the specific CommandHandler
but that seemed even less desirable.