Introduce EntityActions ViewObject
The entity and entity detail view objects are now composed of a entity
action object. Which is used to build the entity actions that can be
performed on them based on the env and status.

The actions themselves have been taken apart ant put in partial templ.
which are included in the two action partials.