Show OIDC confirmation screen after publishing
The modal window is displayed after a successful publication to manage.
This is done only once, after the action was performed the in memory
(session) clone is removed from the session.

The modal window itself is functioning like the type selection modal.

More details: https://www.pivotaltracker.com/story/show/162119902