Update the show/hide rules
If the type of service is set to 'Not an institute' the 'Contract
signed' radio group should be visible. And when institute is selected
the 'SURFconext representative approved' radio group should be visible.

In order to achieve this goal, the service_form.ts was updated severely.

And the RepresentativeApprovedType received a container class to be able
to target it from the service form TS logic.