Load off logic to IssueRepository
The Jira related actions that where performed in the TicketService had a
better fit in the Jira namespace. For that reason, a IssueRepository was
introduced that performs the tasks previously performed by TicketService

For now both the Repo and the Service use the same interface. Which does
not pose any problems yet. This might need to be changed at a later
stage.