Fix routes missing services
The given service of an entity is not set for certain entity routes
which will throw an exception because a service is needed. This was not
an issue before because the active service was set in the session. But
because we now have overview pages with multiple services. We should
add the service to the route to know the active service on forehand.