Fix manage certificate metadata
When a certificate is fetched from manage the start- and end-tag are
stripped. This results in an error when submitting a form again because
the SSL validator thinks it isn't a valid certificate. Therefore
the tags are added in the metadata dto to always return a certificate
with those tags.

https://www.pivotaltracker.com/story/show/163385303