Correctly publish OIDC redirectUris to Manage
The array that is passed along to manage should be a list, not a map.
If the php indexes are updated along the way, by adding/removing items
from the form. The index order might not be sequentially pristine. This
causes JSON encode to generate a map instead of a list.

By resetting the list before publishing, this problem should be remedied

https://www.pivotaltracker.com/story/show/163646662