Provide a custom service overview for admin
The admin user now is faced with a potential 504 error after logging on
to the application. As the service overview page is causing much data
to load for the admin user. In production, over 150 SP's are loaded on
this page including all their entity information.

This does not work out! The easy fix for now is simply rendering a
blank page with a small instruction to the admin user, to use the
service switcher and switch to the desired service that way.

For now the simplest of solutions is chosen (render a different template
if you are an admin user. This could and will be improved upon in an
upcoming development effort.

https://www.pivotaltracker.com/story/show/163691688