Add the service/{serviceId} route
In order to view the service overview as an admin, we want to view the
service overview by service id (the users view them by team id, allowing
them to view multiple services on the screen). To allow interoperability
the new route outputs the same params as the regular user overview. But
signals the template that the admin is watching. Allowing small
distinctions in the rendering of the template. If admin needs grow or
venture away from the user view, we should split these templates. For
now this solution is the most cost efficient.