Streamline service overview route handling
When the empty/no service option is selected, the overview was loaded
for the admin user, resulting in an internal server error. This fix will
display the regular message, instructing the user to use the service
switcher to view service details.

Also, the menu entry for the admin user is updated (renamed) and now
by default points to the service overview of the currenlty active
service.