Do not failonerror with yarn audit
This does not mean we are ignoring vulnerabilities in JS dependencies,
but at the moment we are not yet able to easily fix the 62
vulnerabilities reported. We cannot yet migrate to Jest 24 due to a lack
of TS support. I opened a ticket to fix this once TS support lands for
Jest 24.

https://www.pivotaltracker.com/story/show/164103245