Cover service delete with priv questions in tests
A service with privacy questions is deleted, this is covered in a web
(functional) test. This test is run against a different database which
yields slightly different behavior on the Doctrine ORM side. In other
words, this test does not result in the integrity constraint violation
error. It would, if this tests is run against a database engine that
respects foreign key constraints.