Test that attributes without motivation are shown
The test adds one attribute with motivation and one without to the
SURFnet entity. The test then verifies that both are shown. The one with
the default, not set text. The latter with the expected attribute
motivation that is set on the test entity.

See: https://www.pivotaltracker.com/story/show/165597382