Implement the JsonGeneratorStrategy
This is used by the PublishEntityClients to choose which Json generator
to use. As the saml20/oidc use the existing generator and the oidcng
will  use the OidcngJsonGenerator

An outline of OidcngJsonGenerator was added. This needs testing and
fine tuning according to the needs of the oidc10-rp Manage needs.