Fix the entity delete acl
The voter used on the entity delete endpoint didn't work as expected
for the administrator role because the role was validated against a
string and not a Role.

Also the endpoint didn't take the result into account from the isGranted
method and the isGranted method din't throw an exception either. Both
issues are now fixed.