Escape the wysiwyg translations
The lexik translation bundle  doesn't inpute validate the wysiwyg
entries. Because it's third party software it's hard to get it into
the bundle. Therefore we decided to validate the output.

To do so a wysiwyg filter is introduced and the documentation is
updated accordingly.

Also the ServiceStatusAssembler needed to be updated in order to
escape the json returned to render the service status donut.