Fix loading of empty issue mock file
When loading an empty issue mock file to get a working implementation
without the use of a working jira instance a mock file was used.
If this file was empty an exception was thrown because the data was
then not parsed to an empty array but to null which triggered input
validation. This is now fixed.