Make Jira test mode configurable
We sometimes do not have a working Jira configuration at the ready. In
those cases the test stand-in is a useful fallback measure. Previously
the test and development environment by default used this stand-in. This
has now been changed to become a feature flag.

This flag can be set to true/false in the parameters.yml

The filePath of the test file has also been made into a configuration
option while at it.

https://www.pivotaltracker.com/story/show/167315333