Allow creation of production entities
The formCount solution solved the problem where the two entity type
selection forms are named uniquely. This raised the problem that the
form handler was no longer able to process the production form, as the
formCounter was not raised to '1' when handling the form (as only one
form is created in that scenario).

To fix this (for now), a new form was created when handling a prod form.
In effect setting the formCounter to the correct value for that form.