Show correct client id label on detail view
The oidc entities (oidc and TNG both) where showing their 'entity id' on
the detail view. This is correct from a SPD standpoint, as we borrow the
eid from a saml entity to store the client id. This commit at least
shows the correct label value for the OIDC entites in detail mode.

Thanks @pablothedude for the heads up in:

https://github.com/SURFnet/sp-dashboard/pull/305#issuecomment-530737417