Prevent call to RSCollection function on null
Draft entities without a Rersource Server collection would error into
a 500 status because the Doctrine entity did not have a suitable default
value for the  oidcngResourceServers field. This field should be set
with an ResourceServerCollection instance. But this would be null.
Causing an internal server error when the entity was converted to a
save command.

See bug report in:
https://www.pivotaltracker.com/story/show/168532438