Persist IdP whitelist settings on the form
When working with an entity, the whitelist info is loaded from Manage,
this data would not be saved to the entity in SPD except for in the
explicit IdP whitelist edit screen.

In other situations (entity publish and edit actions) this information
was lost.

By storing the whitelist information on the form in hidden form fields,
we ensure the data is persisted when publishing the metadata changes.

The IdP metadata is serialized to a json encoded string.

https://www.pivotaltracker.com/story/show/168800119