Store the IdP whitelist data
On the entity, when saving a local copy, or when the entity is used
in the domain layer for data transition.

And on the handlers that take care of saving the local copy.

In the publication step, no additional action was required as the white-
list information is already added to the JSON when publishing an entity.

https://www.pivotaltracker.com/story/show/168800119