Persist ACL data on entity edit publication
The IdP ACL data would be overwritten when a published manage entity was
edited. This was the case because the data was still present on the
entity (with the default allow all settings) and the entity is used to
build the Manage Merge Metadata on.

The solution is to load the manage entity prior to publishing it, and
loading the required ACL data on the Entity (that the JSON is based on).
This is only attempted if the entity was found in Manage.

https://www.pivotaltracker.com/story/show/168800119
Alternative to: https://github.com/SURFnet/sp-dashboard/pull/313