Resolve double jQuery version
Both version 3.4.0 and 3.4.1 where included. This leaded to issues with
parsley (who relied on the other version) this resulted in parsley not
being available in the jquery that was used by the rest of the
application.

Solution was to update the resolutions to a version both the application
jq and parsley can live with which turned out to be:

```
"resolutions": {
    "jquery": "3.4.1",
    ...
}
```

Note to future self and other readers. Yarn why can help in identifying
which dependencies use a specific version of a dependency.

Thansk Melvin for helping me in debuggin the frustrating problem!