Allow empty certificate
Previously the array merge function JsonGenerator:242 would try to merge
an undefined variable as the generateCertDataMetadata would not return
an array. This was already fixed on the previous PR in commit 4f320119.
A test was added to verify that this behaviour can not return in the
codebase.

Additionally, a += operator has been changed out for an array_merge
function on line 255. This to prevent issues when the generateLogoMetadata
returns an empty array (would yield a fatal error).