Fix tippy tooltips
CSP rules does not like/restrict using inline css. And tippy wants to
add inline css to style the tooltips. This was resolved by using the
standalone es6 version of tippy.

This version does not by default use the title attribute of the tooltip
to display on the popup. That required a small change to the labels that
are triggering the tooltips.