Prevent overwriting of ARP motivations
The arp motivation would be overwritten when a arp motivations was set
in manage for one of the attributes not managed in SPD. This is solved
by checking for the presence of an attribute motivation. If there is a
non empty value, it is loaded on the attrbibute. Preventing overwriting
that specific attr. motivation.

See:
https://www.pivotaltracker.com/story/show/170868465/comments/214384755