Interpret missing exclude from push correctly
When resetting a client secret on production and the exclude from push
attribute is missing, this indicates we are dealing with a published
attribute. As the service desk not unchecks the coin checkbox, but
simply removes it.

This additional logic covers that (understandable) 'misbehaviour' of
that coin.

https://www.pivotaltracker.com/story/show/171313433/comments/214527925
(cherry picked from commit 3e67fa0604fc7b1131fb1bd2ee4b98f020e5392d)