Validate the privacyPolicyUrl field
The entry should be a valid url. By enabling parsely validation on the
privacy questions form this goal could be achieved. All other fields
needed updated 'required' attributes to prevent them from becoming
required form fields according to Parsley.

In addition, the UrlType is now used, which will prepend https:// when
the protocol is not filled in the form field. Parsley is allowing for
urls without protocol. So this behaviour should always result in a
url with a protocol.

https://www.pivotaltracker.com/story/show/169895852