Improve Url validation logic
The protocol checks now include verification of the double slash after
the protocol. This now fails miswritten domain names like
'https:/foobar.com'

In addition domain names without a TLD are now also rejected. This check
simply tests if a dot is present in the hostname. And not actually
verifies the TLD is a registered/existing top level domain. So the users
mileage may vary.

The previously removed url-validator library was taken as inspiration
for the isIP helper method that was added in this commit. It verifies if
the hostname is a v4 or v6 IP address.

Finally the methods have been cleaned up. Comments have been moved to
the JS doc blocks.