Move the ManageEntity to Domain namespace
This entity will replace the Doctrine entity entity. First it is moved
to the domain layer. Next it will be renamed.

This task also includes optimizing the imports. So some imports might
have been removed due to running this linting task. Might be unrelated
to this change, but was required to correctly sort the imports after
globally search/replacing the references to the moved ManageEnity
classes.