Remove memberOf as managed attribute
The memberOf attribute was defined as attribute but didn't seem to be
managed by SPD. So when pushing the attribute to Manage it was replaced
with an empty value. So therefore the attribute was removed and a
couple of checks for mapping values to getters/setters could be
removed to make the code more explicit and thus better readable.