Prevent method call on null for AllowedIdPs
The merge service would set null for the allowed idps when no specific
value is provided. This is problematic down the road as this triggers a
method call on null in the JSON generator

By providing an instance of the AllowedIdentityProviders VO, set with
the defaults. This issue is resolved