Remove link to detailed entity overview
Prior to this change, there were two entity overviews.

This change simplifies the UI/UX by removing the link to the detailed entity overview from the service overview.  This is a step in the process to removing the detailed entity overview entirely.

See the story in pivotal for more info: https://www.pivotaltracker.com/story/show/165965607