Split entities table in two
Prior to this change, there was only a single table for entities on the service overview.  This despite there being two visual ones.

This change splits the table in two, meaning there are now two tables.  Each has it's own caption and as a result is much more accessible for screen readers.  While working on it i've also cleaned up the styling a tad by removing an unneccesary border (top of the first row) and ensuring the border below the caption title is the full length of the table.

See the story in pivotal for more info: https://www.pivotaltracker.com/story/show/165965607