Update translations for service overview
Prior to this change, the entity tables had titles that were very domain language / tech oriented.  The title of the page also did not reflect the content, as it still assumed the old situation where multiple services could be on this page.

This change ensures those titles are more readable & that the title reflects the content of the page.

See the story in pivotal for more info: https://www.pivotaltracker.com/story/show/165965607