Move add entity button to service overview
Prior to this change, the add entity button was located on the detailed entity overview.

This change ensures the buttons are moved to the service overview.  They are now also located behind the table titles.  The text for the title has also changed to reflect the kind of entity being added.
Note that the cancel button doesn't work at this time.  This will be fixed when taking up https://www.pivotaltracker.com/story/show/165599582

See the story in pivotal for more info: https://www.pivotaltracker.com/story/show/165965607