Increase default font-sizes
Prior to this change, the font-sizes were very small.  This makes the page very hard to read and creates accessibility problems.

This change ensures the font-sizes on the service overview are increased to account for readability and accessibility.  It also removes the primary contact column from the entity tables because the bigger font-sizes mean they need less info in them to maintain an easy overview.
While at it some semantic improvements were made as well.

See the story in pivotal for more info: https://www.pivotaltracker.com/story/show/165965607