Change back link destination on entity detail
Prior to this change, the back-link returned to the entity detail overview.

This change ensures the back-link goes to the relevant service overview.  The reason for the change is that the entity detail overview will be removed shortly.  It was also possible to reach the entity detail from both the service overview and from the entity detail overview.  Given that is no longer possible (the origin is always the service overview), the back link now makes more sense.

See the story in pivotal for more info: https://www.pivotaltracker.com/story/show/165965607