Copy entitylist-specific code to service controller
Prior to this change, everything was handled in the entityListController.

This change ensures two variables are correctly passed to the service overview.  It also ensures the behaviour they enabled in the entity detail overview is correctly copied to the service overview & that all associated controllers are changed as well.

See the story in pivotal for more info: https://www.pivotaltracker.com/story/show/165965607